Bug fixed: circular import. (#305)
